some consistency tweaks
make the expires value passed to the store_auth_code callback also
a validity period in seconds, so it is consistent with the other
store_access_token callback. update tests and examples to reflect
this change

remove the client id from the generated auth code / access token as
this could classify as data leakage (not really given the original
request for an auth code is a GET, nonetheless...)

slight refactoring to make setting of the _store_auth_code callback
consistent

when mapping $self->param use // to set undef if not defined. this
seems unnecessary but without it tests fail on my machine. version
of Mojolicious perhaps? (5.38, perl 5.18.1)

remove some debug, bump VERSION, Changes