document Path gotcha in flash (login_resource_owner)
w/r/t c782678074986cff61581815c74e7b1f7ba882d6 - the bug in my real
code was that the flash was set on the /oauth/ Path so when i go to
the /login route i can't access that flash value (different Path).

the fix was to alias /oauth/login to /login so i *can* access the
flash value set in login_resource_owner. document this potential
gotcha in both perldoc and examples, and update tests to use the
/oauth/login and /oauth/confirm_scopes routes so it is clearer