ref #2 make jwt_secret optional
so we have back compat and also the option to either just JWTs or
not use JWTs. add tests and examples for using JWTs, update pod.

TODO: update the module's verify_auth_code and verify_access_token
to do what the example oauth2_server_simple_jwt.pl does when the
jwt_secret is set - this will allow the plugin to be used in a
persistent and multiproc env (but will lack revoking capabilities)