Accept error_description from auth methods
Returning `error` in redirect URL when authorization failed in most
cases is not expressive enough and does not give enough context to the
caller what excatly went wrong.

As per RFC we can also add `error_description` (which is already done in
the plugin for couple of flows) that can help caller to pinpoint the
issue. Change here is to expect potential 3rd retun value from Mojo
OAuthv2 plugin and add the error to redirect URL.

RFC:
https://datatracker.ietf.org/doc/html/rfc6749#section-5.2