resolve #28 - allow client credentials in the body
looking at the original RFC I see this:
https://datatracker.ietf.org/doc/html/rfc6749#section-2.3.1

    "Including the client credentials in the request-body using the two
    parameters is NOT RECOMMENDED and SHOULD be limited to clients unable
    to directly utilize the HTTP Basic authentication scheme (or other
    password-based HTTP authentication schemes)."

so it seems using the credentials in the body isn't really recommended.

however it *is* supported, so support it. update test coverage to
exercise this as well as fixing a bug revealed in a test by implementing
it.

bump VERSION and Changes for CPAN release